Use Terraform to generate local kubeconfig file
Instead of using Ansible and a provisioner to generate the local
kubeconfig file, we now use a variety of Terraform data sources and
resources to generate the kubeconfig using a template and variables.

This fits in better with the new Argo CD workflow, and allows Terraform
to manage the kubeconfig file like any other resource.