feat(k8s): add ConfigMap vars via Terraform
Adds the ability to define a map of key/value pairs in Terraform, which are then created as a ConfigMap in the K8s cluster. Flux can then use this ConfigMap for variable substitution when applying manifests.

This is a copy of the existing feature to seed Secrets, using the same process.

MR: https://gitlab.com/ralgar/homelab/-/merge_requests/29