chore: change to permissive license
Change to the BSD 2-clause "Simplified" or "FreeBSD" license. This license is simple and permissive, and is better suited for a project of this nature than the previously used GPL 3 license.

https://choosealicense.com/licenses/bsd-2-clause/

NOTE: The markdownlint CI job has also been updated to use a different container, since the previous implementation seemed to have no convenient way of disabling rules on a global or per-file basis.

MR: https://gitlab.com/ralgar/homelab/-/merge_requests/33