fix(k8s): externally provision media volume
The media volume must be persistent, so this MR changes it to be externally provisioned (through Terraform). This way, we can rebuild clusters as needed, while retaining the data from this large volume which isn't feasible to backup.

This MR also temporarily scales the cluster down to 1 xlarge worker node, since multiattach is not working correctly for the volume.

MR: https://gitlab.com/ralgar/homelab/-/merge_requests/58